Decodes whole functions at a time now. Functions are now indexed in terms of (pc, bytes hash), and (page range hash, pc), where I discovered that doing whole functions at a time led to more than one function having the same (hash1,hash2) values (this was a surprise).

+++ vmill/BC/Trace.h
@@ -26,7 +26,7 @@ class Module;
 }  // namespace llvm
 namespace vmill {
 
-using TraceHashBaseType = uint32_t;
+using TraceHashBaseType = uint64_t;
 enum class TraceHash : TraceHashBaseType;
 enum class PC : uint64_t;
 enum class CodeVersion : uint64_t;
